g4-common: thermal: massive thermal change (2/2)
... part 1/2 = kernel.

TL;DR

Bringing back thermal-engine, fix and optimize everything around thermal based
frequency throttling and ensure device does not get HOT like HELL anymore.
The kernel handles hotplug of all cores if needed, while frequencies will be handled
by the thermal-engine service.

Details:

 - resurrect thermal-engine service (+ fix init)
 - fixing all useless and non-working thermal profiles (e.g. SS-CPU3 etc)
 - add and optimize cluster frequency handling
 - fix GPU monitor and optimizing frequency handling
 - fix Battery monitor
 - fix CPU_SENSORS (omg..)
 - fix all samplings

Explained:

The thermal-engine should (IMHO) not be used for hotplugging the cores. The kernel
is way more effective on that and the configuration there is optimized for the
LG G4.

On the other site handling frequency throttling on kernel level is not very flexible
and so you need the thermal-engine service for monitoring and handling that correctly.

If you still wanna use thermal-engine to handle hotplugging ensure you check the kernel
level first to avoid ping-pong between.

Signed-off-by: steadfasterX <steadfasterX@gmail.com>